Adds 'global' configuration overrides, useful for CI
This looks for APP_CONFIG_EXTEND or APP_CONFIG_CI for valid
configuration objects, and assigns those variables on top of
the existing configuration.

This isn't really meant as a production variable (use APP_CONFIG),
but it is very very useful for CI.

It is smart, in that it assigns itself as a secret if that secret
was already present, so to not have two properties of the same name.